Added ability to modify file load order by appending number to filename.
The natural sorting algorithm of PHP used to do a sloppy job, and didn't
order correctly when fractions were used. The new algorithm first tries to
compare by number in the beginning of the file basename, and compares
fractions correctly. If no number used, fall back to natural sorting.
This allows the ability to always insert a file in between any 2 files in
the load order by specifying a more precise fraction.